Code: using enum class for FormatType

+++ PowerEditor/src/ScitillaComponent/ScintillaEditView.cpp
@@ -1577,7 +1577,7 @@ void ScintillaEditView::bufferUpdated(Buffer * buffer, int mask)
 
 		if (mask & BufferChangeFormat)
 		{
-			execute(SCI_SETEOLMODE, _currentBuffer->getFormat());
+			execute(SCI_SETEOLMODE, static_cast<int>(_currentBuffer->getFormat()));
 		}
 		if (mask & BufferChangeReadonly)
 		{
